+++ FinanceMeKit/FinanceMeKit/UI/Common/TransactionViewModel.swift
@@ -2,7 +2,7 @@ struct TransactionViewModel {
     let narrative: String
     let amountViewModel: AmountViewModel
 
-    init(narrative: String, amount: Decimal, signs: [AmountViewModel.Sign] = [.plus]) {
+    init(narrative: String, amount: Double, signs: [AmountViewModel.Sign] = [.plus]) {
         self.narrative = narrative
         self.amountViewModel = AmountViewModel(value: amount, signs: signs)
     }
